+++ src/adaptor/preact.ts
@@ -1,0 +1,1 @@
+export { useEffect, useState } from 'preact/hooks'